Add files via upload
It is a type of input device that detects and responds to multiple simultaneous touches or gestures on its surface. It relies on the principles of capacitive sensing, which involves measuring changes in capacitance to detect the presence and location of touch inputs. So this is done by Capacitive sensors. A capacitive sensor is an electronic device that can detect solid or liquid targets without physical contact .Changes in capacitance, which is the ability to store an electric charge, between two conductive surfaces. When a finger or other conductive object approaches or touches the graphene layer, it modifies the capacitance of the graphene structure. The change in capacitance caused by the touch input can be measured by applying an electrical signal to the graphene layer and monitoring the resulting changes in electrical properties and that resulting changes should be accurate .
Now we are making it more user friendly by using Arduino UNO as a controller ,due to this we do not need to apply a separate electrical supply to this. We observe this results by using a opensource platform called Pygame. Where we show the same replica of grid in digital format. And this makes it easy and fun to understand the changes.